Update dependencies_proper_format.csv
Hi Serge,

I have updated this document. 

here are my comments, questions:

States of character 35 (presence of labral apodemes) should be dependent on character 25 (presence of labrum) and inapplicable when labrum is absent, however, in those taxa where the labrum is present state 0 (labral apodemes absent) has ben scored.

Character 56 (paraglosal basal setation) depends on char 55 (Paraglossae) since state 4 is "paraglossae totally reduced" so should not have any setation. However, other than state 4 have been scored for taxa here paraglossae are absent. Also, I don't really know how to enter these to the list since more than one "present" states (on which the next character is based) present. Again, they did a mistake in their scoring, so perhaps this character should not be involved in our analysis.

The same problem in character 63 (sensillum on third maxillary palpomere) vs character 61 (number of maxillary palpomeres). Multiple states of 61 have maxillary palpomere present.

Same problems with 95 (cervical apodeme presence) and 96 (cervical apodeme configuration). Multiple present character states.

Character 67 (galea division), I guess, is dependent on character 66 (galea shape) which has a state "absent" I guess meaning "galea absent". However, even in taxa with "galea absent" (e.g. Tremex or Sirex) character 67 is scored (unipartite).

Character 122 should be inapplicable if probasitarsal notch is absent, since probasitarsal comb is always in probasitarsal notch so if notch is absent, comb is absent, too. ch 122 is scored as absent if char 121 is scored absent.

Character 126 is misscored at Ceraphron (character 124 is "0" so the character 126 should be inapplicable)

Character 128 should be inapplicable if character 127 is "0", but this is not the case. Multiple misscoring with character128.

Multiple presence states in character 131 (character 132 is dependent on presence states of character 131)

Multiple present character states in character 135 (character  136 is dependent on present states of character 135)

character 191 should be inapplicable if character 190 has state "0" (this is off course not the case)

Multiple presence states in character 250 (character 252 is dependent on presence states of character 250)

Character 88 is applicable if character 85 OR character 87 are present.

Character 89 is applicable if character 85 OR character 87 are present.

Character 174 is applicable only if character 173 AND character 171 are present

Multiple presence states in character 95 (character 96 is dependent on presence states of character 95)